requirements.txt
absl-py==0.9.0
appnope==0.1.0
argh==0.26.2
asgiref==3.2.3
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
coremltools==3.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Django==3.0
docker==4.1.0
docutils==0.15.2
entrypoints==0.3
fire==0.2.1
gast==0.2.2
gitdb2==2.0.6
GitPython==3.0.5
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageai==2.1.5
imageio==2.6.1
importlib-metadata==1.3.0
imutils==0.5.3
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras2onnx==1.6.0
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.17.4
oauthlib==3.1.0
onnx==1.6.0
onnxconverter-common==1.6.0
onnxmltools==1.6.0
opt-einsum==3.1.0
pandocfilters==1.4.2
parso==0.5.2
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.5.2
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22
scipy==1.3.3
Send2Trash==1.5.0
six==1.13.0
skl2onnx==1.6.0
smmap2==2.0.5
sqlparse==0.3.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-hub==0.7.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf2onnx==1.5.3
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
traitlets==4.3.3
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
